correction of comments
