+++ src/app/create-todo-form/create-todo-form.component.html
@@ -7,11 +7,6 @@
      <input matInput type="text" formControlName="title" />
    </mat-form-field>
    <mat-error
-     *ngIf="
-       formTodo.controls['title'].invalid &&
-       (formTodo.controls['title'].dirty ||
-         formTodo.controls['title'].touched)
-     "
    >
      Недопустимая длина символов
    </mat-error>
@@ -21,11 +16,6 @@
      <input matInput type="text" formControlName="userId" />
    </mat-form-field>
    <mat-error
-     *ngIf="
-       formTodo.controls['userId'].invalid &&
-       (formTodo.controls['userId'].dirty ||
-         formTodo.controls['userId'].touched)
-     "
    >
      Недопустимая длина символов
    </mat-error>
@@ -35,17 +25,12 @@
      <input matInput type="text" formControlName="completed" />
    </mat-form-field>
    <mat-error
-     *ngIf="
-       formTodo.controls['completed'].invalid &&
-       (formTodo.controls['completed'].dirty ||
-         formTodo.controls['completed'].touched)
-     "
    >
      Поле заполнено некорректно
    </mat-error>
    <div class="create-button-position">
     <mat-card-actions>
-     <button mat-raised-button (click)="onNoClick()">Закрыть</button>
+     <button mat-raised-button [mat-dialog-close]="">Закрыть</button>
      <button mat-raised-button type="submit" [mat-dialog-close]="submitForm()" [disabled]="
      formTodo.invalid">Создать задачу</button>
     </mat-card-actions>
