+++ src/app/dialogs/user-dialogs/create-user-dialog/create-user-dialog.component.html
@@ -8,10 +8,6 @@
        <input matInput type="text" formControlName="name" />
      </mat-form-field>
      <mat-error
-       *ngIf="
-         form.controls['name'].invalid &&
-         (form.controls['name'].dirty || form.controls['name'].touched)
-       "
      >
        Слишком короткое имя
      </mat-error>
@@ -21,10 +17,6 @@
        <input matInput type="text" formControlName="email" />
      </mat-form-field>
      <mat-error
-       *ngIf="
-         form.controls['email'].invalid &&
-         (form.controls['email'].dirty || form.controls['email'].touched)
-       "
      >
        Email должен содержать больше 3-х символов и {{ "@" }}
      </mat-error>
@@ -34,10 +26,6 @@
        <input matInput type="text" formControlName="website" />
      </mat-form-field>
      <mat-error
-       *ngIf="
-         form.controls['website'].invalid &&
-         (form.controls['website'].dirty || form.controls['website'].touched)
-       "
      >
        Поле заполнено некорректно
      </mat-error>
@@ -47,16 +35,11 @@
        <input matInput type="text" formControlName="name" />
      </mat-form-field>
      <mat-error
-       *ngIf="
-         form.controls['name'].invalid &&
-         (form.controls['name'].dirty ||
-           form.controls['name'].touched)
-       "
      >
        Поле заполнено некорректно
      </mat-error>
      <div class="create-button-position">
-      <button mat-button (click)="onNoClick()">Закрыть</button>
+      <button mat-button [mat-dialog-close]>Закрыть</button>
        <button type="submit" (click)="submitForm()" mat-raised-button [disabled]="form.invalid">Создать</button>
      </div>
    </form>
