+++ src/app/dialogs/user-dialogs/delete-user-dialog/delete-user-dialog.component.html
@@ -4,7 +4,7 @@
   <p>Вы действительно хотите удалить пользователя?</p>
 </div>
 <div mat-dialog-actions class="button-position">
-  <button mat-raised-button (click)="cancelDelete()">Нет</button>
+  <button mat-raised-button [mat-dialog-close]>Нет</button>
   <button mat-raised-button (click)="confirmDelete()">Да</button>
 </div>
 </mat-card>
