+++ src/app/dialogs/user-dialogs/delete-user-dialog/delete-user-dialog.component.ts
@@ -1,18 +1,17 @@
-import { Component } from '@angular/core';
-import { MatDialogRef } from '@angular/material/dialog';
+import { Component, inject } from '@angular/core';
+import { MatDialogClose, MatDialogRef } from '@angular/material/dialog';
 import { MatCardModule } from '@angular/material/card';
 import { MatButtonModule } from '@angular/material/button';
 
 @Component({
   selector: 'app-delete-user-dialog',
   standalone: true,
-  imports: [MatCardModule, MatButtonModule],
+  imports: [MatCardModule, MatButtonModule, MatDialogClose],
   templateUrl: './delete-user-dialog.component.html',
   styleUrl: './delete-user-dialog.component.scss',
 })
 export class DeleteUserDialogComponent {
-  constructor(public dialogRef: MatDialogRef<DeleteUserDialogComponent>) {}
-
+  private dialogRef = inject(MatDialogRef<DeleteUserDialogComponent>);
   public confirmDelete(): void {
     this.dialogRef.close(true);
   }
