+++ src/app/dialogs/user-dialogs/edit-user-dialog/edit-user-dialog.component.html
@@ -6,10 +6,6 @@
    <input matInput type="text" formControlName="name" />
  </mat-form-field>
  <mat-error
-   *ngIf="
-     form.controls['name'].invalid &&
-     (form.controls['name'].dirty || form.controls['name'].touched)
-   "
  >
    Слишком короткое имя
  </mat-error>
@@ -19,10 +15,6 @@
    <input matInput type="text" formControlName="email"
  /></mat-form-field>
  <mat-error
-   *ngIf="
-     form.controls['email'].invalid &&
-     (form.controls['email'].dirty || form.controls['email'].touched)
-   "
  >
    Email должен содержать больше 3-х символов и {{ "@" }}
  </mat-error>
@@ -32,10 +24,6 @@
    <input matInput type="text" formControlName="website"
  /></mat-form-field>
  <mat-error
-   *ngIf="
-     form.controls['website'].invalid &&
-     (form.controls['website'].dirty || form.controls['website'].touched)
-   "
  >
    Поле заполнено некорректно
  </mat-error>
@@ -45,16 +33,11 @@
    <input matInput type="text" formControlName="name" />
  </mat-form-field>
  <mat-error
-   *ngIf="
-     form.controls['name'].invalid &&
-     (form.controls['name'].dirty ||
-       form.controls['name'].touched)
-   "
  >
    Поле заполнено некорректно
  </mat-error>
  <div class="button-position" >
-  <button mat-button (click)="onNoClick()">Закрыть</button>
+  <button mat-button [mat-dialog-close]>Закрыть</button>
   <button mat-raised-button type="submit" [disabled]="form.invalid" [mat-dialog-close]="userWithUpdatedFields">Редактировать</button>
  </div>
 </form>
