+++ src/app/todos-list/todo-card/todo-card.component.html
@@ -10,7 +10,7 @@
       <button mat-raised-button (click)="openDeleteTodoDialog()">
        Удалить задачу
       </button>
-      <button mat-raised-button (click)="openDialog()">
+      <button mat-raised-button (click)="openEditTodoDialog()">
        Редактировать задачу
       </button>
     </div>
