+++ src/app/todos-list/todos-list.component.ts
@@ -42,9 +42,9 @@ export class TodosListComponent {
   }
 
   editTodo(todo: Todo) {
-   this.todosService.editTodo({
-    ...todo,
-   })
+    this.todosService.editTodo({
+      ...todo,
+    });
   }
 
   public createTodo(formData: Todo): void {
