+++ src/app/validator.ts
@@ -1,0 +1,11 @@
+import { AbstractControl, ValidatorFn, ValidationErrors } from '@angular/forms';
+
+export function completedValidator(): ValidatorFn {
+  return (control: AbstractControl): ValidationErrors | null => {
+    const value = control.value?.trim().toLowerCase();
+    if (value === 'да' || value === 'нет') {
+      return null;
+    }
+    return { invalidCompleted: true };
+  };
+}
